requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Flask==2.2.3
Flask-API==3.0.post1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-jsontools==0.1.7
flask-marshmallow==0.15.0
Flask-MySQLdb==1.0.1
Flask-QRcode==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
future==0.18.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
pyotp==2.8.0
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
sa-filters==1.2.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
